Test to list all Assignments for each Class (#2)
The test checks whether the returned class ids are valid.
The test checks that a list of assignments can be fetched given a class
id.
The test checks that the assignments returned by each class are not
identical.

Changed the nullibility of some fields of assignment based on errors
when
validating a larger set of assignments than just todo.